+++ test/flows/auth.flows.ts
@@ -1,4 +1,4 @@
-import { Auth } from "../../pages/auth.pages";
+import { Auth } from "../pages/auth.pages";
 
 export async function register(username: string, password: string, auth: Auth) {
